refactor: rearrange entities files
